Workaround for being able to start with `Enter_menu_start()`
The issue is that some of the drawing passes are drawn even when
menu is active, but the game play is not yet started.

We might want to properly initialize gameplay and avoid the runtime
check, but it will require a bit fuller integration of cocktail
manager. Committing something temporary for the time being.