Attempt crash hot fix
Crash happens in unclear conditions when pressing B and the frog speaks
The frog string is a table and not a string, but I don't know why yet.

(cherry picked from commit 15e8d79d3bea11af48d5727b1b75ab733bad7ea4)

Integration note:
The crash turned out to be due to the fire_reminders being a table.
This was discovered and fixed in 92e8b7b31
That commit won't be integrated in the jam version branch as it already has significant restructuring to the frog dialog. The jam branch gets only the safeguard for crashing.