Tidy up animation resources, reduce runtime memory usage by 50KB
Turns out that animation loops *copy* the image table they're given :/
Workaround by passing in the imagetables directly when creating the animation timers, this is suboptimal as now there is no separation between resource loading and animation state, but in this game there is no sprite instantiation, so it likely won't matter.
Note: didn't do the same for cocktails yet as it wasn't trivial.

Runtime memory usage during gameplay is no 2.30MB -> 2.25MB