Smarter and more helpful talk reminders
If B is never pressed before the Frog has automated messages.
These remain helpful and as little annoying as possible.
Frog always says what he thinks. Struggles speed up the frogs response.
The frog never comments on fire or stirring by himself,
unless it's important for finishing the cocktail.
If the cocktail is done, the frog will give a hint to end it.
Various other small fixes and cleanups.