Added sticker slap and mastered sticker
When first completing a cocktail or perfecting it, a hand slap anim
plays when returning to the mission selection.
This is done with a new menu state to disable inputs for a short time.
New gameplay states are added to keep track of the completion of the
recent cocktail.
Also some additional cleanup and added comments.