Small updates to frog sound logic and comments
I set up the burp and drinking sounds correctly for future implementation.
The eyelick sound is set to loop and only restart if it's not playing.
The sounds are all stopped before the new one is set to start playing.
Sounds are only stopped if they actually play (to prevent bugs).