Fixed bug for recipe struggles
When reaching struggle steps, the timer would endlessly restart
the struggle flag and animation.
To fix it I removed the timout timer, which isn't aactually needed,
since the conditions are based on the recipe steps instead of a timer.
Talking to the frog or making more steps will resolve the frog state.