Initial attempt at fixing rune travel jumping
BUG: When stirring half way and dropping different ingredients,
the average rune position changes and snaps.
This gives the wrong impression of how runes move and work.

DESIRED: The rune positions is fixed in place until stirring further.
THis can be acheived by calculating them together from all dropped
rune changes and how much each of them have been stirred in since they
have been added.

Initial commit:
I refactored the code to include a rune_count_current gameplay state.
This one is calculated for the rune travel visuals.
It is calculated from a CURRENT_DROPS dictionary of rune changes from each drop,
and stir factor for each drop as a factor.
All of them added together with a stir factor of 1 should have the same
result as the default rune_count gameplay state with a global stir factor
of 1.

Current state:
Some fixes still needed to make the math work.
The stir factors sync up correctly. Eventually the global stir factor
should be derived from the addition of the CURRENT_DOPS stir factors.
But the rune count is still not correct. One reason is that it's not
properly clamped.