Updated debug code
The rune_count and rune_count_current are perfectly in synch.
The remaining issue is that the stir factor gets more out of sync
the more half stirred ingredients are added.
The CURRENT_DROPS stir factor is more accurate but more nuanced.
The best way to fix this is to derive the global stir factor from the
CURRENT_DROPS table entirely.