Split menu and gameplay code paths
Originally the code was made to allow overlaying of menu screens (pause, gameover, inventory, wtv)
on top of a running game, but in this game it's either in the menu or not, in which case we don't
want to be updating and drawing the gameplay under the opaque menu for simplicity reasons.

Code is now updated to either update and draw one or the other.
The gameplay uses Playdate's Sprite system, the menu doesn't.
Because the sprite system isn't running on the menu, we don't need to add/remove sprites anymore.
To be considered: remove the sprite system also from the gameplay? We barely use the features.

This fixes the 15FPS performance when going back to the cocktail menu after playing.