Recipe drawing: fixup scrolling for gameplay version
Cleanup interface for the scrolling (was lowercase global variables). Reduced calculations on input.

- Rename global end_recipe_y to RECIPE_SCROLL (hopefully can be re-used by the menu as well).
- Added new global for the scroll range instead of calculating every frame.
Calc is done in function close to the layout, easier to spot it going off sync.
Pre-calculating all the layout onto a structure doesn't seem worth it.
- Remove global scroll_speed = 1.8. It was used only as local constant (and shadowed in menu), but then always multiplied by 2.
As this is an increment in pixels, I set it to 4.

- Fix recipe layout going on a half pixel due to line_height/2
? Crankticks doesn't need an initial call after the animation?