Liquid bubbles: refactor and minor improvements
Refactored the bubble drawing code with the goal of improving performance, it takes 3ms.
Unfortunately, it took a lot of time and there's no gains.

What takes time:
- 50% for drawing, roughly proportional to the size of the sprite. (drops cost more than bubbles).
- Using an image table instead of an image is 20% of the draw cost, not sure how to work around that.
- 15% lifetime and frame math, e.g. floor
- 20% using sin (plus another sin for drops which makes them costlier)
- Image mirroring, speed squish and the spawn code including randoms have a minor impact.
An approximation of sin could be worth it, but not worth it for my own fun.
Pre-calculating the ellipse paths and then only sampling according to angle would help, but again
I find it a bother and the squish factor makes them dynamic. They're also repeated or very similar
due to the amplitude math.

On the up side, the code is now hopefully understandable and there's no hacks.

CBB: this code is updating and drawing, but I don't see a reason to spend the time correcting now.

TODO: trim white space in bubble and drop sprites for perf.