Sprite system tweaks
Disable collisions which we don't use and set things to always redraw to avoid the occlusion system
which seems to have little benefit.
It could be better to not use sprites at all, since we only use the draw order, move to, set image and update.
Ideally we'd have a gameplay draw functioning explicitly drawing everything in Z order and more directly keep
track of ingredients instead of having anything from the sprite system, but it's still a bit of work to type.