Code cleanup for menus
- Moved timers used for the menus that were under "gameplay" timers.
This is no longer misleading and fixes the problem with a variable being global (from the times where there was still hope for that).
- Renamed menu transition functions according to convention.
- Moved calculations and variables to inside the conditions where they're used.

I wanted to fix transitions being done on timers and hard to parse logic, but it's very very hard to parse with little benefit from cleaning it up unless we have to.