Fix frog not licking his eyeballs when he should - significant changes
Repro: start snailiva, add perfume drops until heart glows.
-> Frog does nothing until something else happens e.g. after he speaks.

Logic for Froggo:Lick_eyeballs() was not being run because it was dependant on CHECK_IF_DELICIOUS
That variable has complex logic to be on and I see no reason to not always check if the frog should react to the cocktail being good.

Removed CHECK_IF_DELICIOUS.
Fallout: adjusting the logic and fixing the eyeball licking from not changing the frog's idle state
could have all kinds of unpredictable side effects due to the complex conditional logic with timers for the frog and gameplay.
I couldn't do a simple fix, had to re-write quite a bit for it to make sense and can't speak for side effects.