Fix frog teaching about fire when fire is already on
The issue was that I started the fire (which doesn't mark it as learned),
then started adding ingredients and at some point the frog times in to explain fire, although it's perfectly on.
That's because I didn't make new flames after shaking in the first ingredient, but it's still visibly on.
If it would have been off already, I don't think it's unreasonable that he teaches (instead of reminds) to have it on.

Repro steps:
- start Silkini (or wtv)
- start a big fire
- add 4 spiders, stir
- wait a bit (all tutorials should be completed and fire visibly on)
-> frog should say something else than teaching about fire.