Undo fix for frog explaining about fire when fire is on
Fix added in 1e8f42d
Undone because it was jointly considered more important to not lose a beginner player than annoy an experienced one.
It is possible that a beginner activated the fire with their thumbs when starting the game and if they learned
how to shake an ingredient fast enough before the fire goes out, they'll never learn the fire.